...G...... [ZBX-8127] changed variable type
git-svn-id: svn://svn.zabbix.com/branches/dev@45335 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ ZBX-8127/src/libs/zbxsysinfo/hpux/hostname.c
@@ -26,7 +26,7 @@ ZBX_METRIC	parameter_hostname =
 int	SYSTEM_HOSTNAME(AGENT_REQUEST *request, AGENT_RESULT *result)
 {
 	char	*hostname;
-	int 	hostbufsize = 0;
+	long 	hostbufsize = 0;
 
 #ifdef _SC_HOST_NAME_MAX
 	hostbufsize = sysconf(_SC_HOST_NAME_MAX) + 1;
